+++ CHANGES.md
@@ -8,6 +8,7 @@
 - SQLAlchemy: Clean up and refactor SQLAlchemy polyfills
   to `cratedb_toolkit.util.sqlalchemy`
 - CFR: Build as a self-contained program using PyInstaller
+- CFR: Publish self-contained application bundle to GitHub Workflow Artifacts
 
 ## 2024/06/18 v0.0.14
 - Add `ctk cfr` and `ctk wtf` diagnostics programs