+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 
 ## Unreleased
+
+## 2024/07/08 v0.0.15
 - IO: Added the `if-exists` query parameter by updating to influxio 0.4.0.
 - Rockset: Added CrateDB Rockset Adapter, a HTTP API emulation layer
 - MongoDB: Added adapter amalgamating PyMongo to use CrateDB as backend